feat: refactor DashboardPage to retrieve accessToken using getAccessToken function
