+++ src/app/dashboard/page.tsx
@@ -1,15 +1,9 @@
-import { jwtVerify } from 'jose';
 import { GetAccountResponse } from '@/shared/types/Accounts.types'
-import { cookies } from 'next/headers'
 import { Card } from "flowbite-react";
-import { COOKIE_SESSION_KEY } from '@/shared/constants/Global.constants';
+import { getAccessToken } from '@/shared/lib/auth';
 
 export default async function DashboardPage () {
-  const secretKey = process.env.SESSION_SECRET_KEY!
-  const session = cookies().get(COOKIE_SESSION_KEY)?.value ?? ''
-  const encodedKey = new TextEncoder().encode(secretKey)
-  const jwtDecoded = await jwtVerify(session, encodedKey)
-  const accessToken = jwtDecoded?.payload?.accessToken as string
+  const accessToken = await getAccessToken()
   const res = await fetch('http://localhost:6006/account-actions', {
     headers: {
       'Authorization': `Bearer ${accessToken}`
